Move farms & status logic to separate modules.
